wrote test for checking load todos from the service

+++ src/app/2-todos/todos.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { TodoService } from './todo.service'
+import { TodoService } from './todo.service';
 
 @Component({
   selector: 'app-todos',
@@ -8,23 +8,22 @@ import { TodoService } from './todo.service'
 })
 export class TodosComponent implements OnInit {
   todos: any[] = [];
-  message; 
+  message;
 
   constructor(private service: TodoService) {}
 
-  ngOnInit() { 
-    this.service.getTodos().subscribe(t => this.todos = t);
+  ngOnInit() {
+    this.service.getTodos().subscribe(t => (this.todos = t));
   }
 
-  add() { 
+  add() {
     var newTodo = { title: '... ' };
-    this.service.add(newTodo).subscribe(
-      t => this.todos.push(t),
-      err => this.message = err);
+    this.service
+      .add(newTodo)
+      .subscribe(t => this.todos.push(t), err => (this.message = err));
   }
 
   delete(id) {
-    if (confirm('Are you sure?'))
-      this.service.delete(id).subscribe();
-  } 
+    if (confirm('Are you sure?')) this.service.delete(id).subscribe();
+  }
 }
